BREAKING: Rename `SlotV2` to `Slot` and `SlotableV2` to `Slotable`.
