+++ performance/components/slots_component.html.erb
