+++ performance/components/slots_component/item_component.html.erb
