+++ test/sandbox/app/components/inherited_slots_component.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class InheritedSlotsComponent < SlotsComponent
+end
