--- test/sandbox/app/components/inherited_slots_v2_component.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-class InheritedSlotsV2Component < SlotsV2Component
-end
