+++ test/sandbox/app/components/nested/slots_component.html.erb
