+++ test/sandbox/app/components/nested/slots_component/another_component.html.erb
