+++ test/sandbox/app/components/nested/slots_component/my_highlight_component.html.erb
