+++ test/sandbox/app/components/slots_before_render_component.html.erb
