+++ test/sandbox/app/components/slots_before_render_component/my_greeting_component.html.erb
