+++ test/sandbox/app/components/slots_before_render_component/my_title_component.html.erb
