+++ test/sandbox/app/components/slots_block_component.html.erb
@@ -1,4 +1,4 @@
-<%= render SlotsV2Component.new(classes: "mt-4") do |component| %>
+<%= render SlotsComponent.new(classes: "mt-4") do |component| %>
   <% component.with_title do %>
     This is my title!
   <% end %>
