+++ test/sandbox/app/components/slots_block_component.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class SlotsBlockComponent < ViewComponent::Base
+end
