+++ test/sandbox/app/components/slots_component.html.erb
