+++ test/sandbox/app/components/slots_component/my_highlight_component.html.erb
