+++ test/sandbox/app/components/slots_delegate_component.html.erb
