+++ test/sandbox/app/components/slots_delegate_component.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class SlotsDelegateComponent < ViewComponent::Base
+  renders_many :items, SlotsComponent::MyHighlightComponent
+end
