--- test/sandbox/app/components/slots_v2_block_component.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-class SlotsV2BlockComponent < ViewComponent::Base
-end
