--- test/sandbox/app/components/slots_v2_delegate_component.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-class SlotsV2DelegateComponent < ViewComponent::Base
-  renders_many :items, SlotsV2Component::MyHighlightComponent
-end
