+++ test/sandbox/app/components/slots_with_empty_lambda_component.html.erb
