+++ test/sandbox/app/components/slots_with_pos_arg_component.html.erb
