+++ test/sandbox/app/views/integration_examples/slots.html.erb
@@ -1,4 +1,4 @@
-<%= render(SlotsV2Component.new) do |component| %>
+<%= render(SlotsComponent.new) do |component| %>
   <% component.with_title do %>
     <p>This is my title!</p>
   <% end %>
