+++ test/sandbox/test/view_component_system_test.rb
@@ -34,7 +34,7 @@ class ViewComponentSystemTest < ViewComponent::SystemTestCase
   end
 
   def test_components_with_slots
-    with_rendered_component_path(render_inline(SlotsV2Component.new) do |component|
+    with_rendered_component_path(render_inline(SlotsComponent.new) do |component|
       component.with_title do
         "This is my title!"
       end
